move events js to stimulus controller
