+++ app/assets/javascripts/global-modules/show_event.js
@@ -1,10 +1,10 @@
 WebsiteOne.define('showEvent', function() {
   function showEvent () {
-    this.showUserTimeZone = function () {
-        if ($("#local_time").length) {
-            $('#local_time').append("&nbsp;" + WebsiteOne.timeZoneUtilities.detectUserTimeZone());
-        }
-    };
+    // this.showUserTimeZone = function () {
+    //     // if ($("#local_time").length) {
+    //     //     $('#local_time').append("&nbsp;" + WebsiteOne.timeZoneUtilities.detectUserTimeZone());
+    //     // }
+    // };
 
     this.toggleDropdown = function() {
       $(".dropdown-menu a").click(function() {
